+++ lib/src/velocity_x_extensions.dart
@@ -11,6 +11,7 @@
  * limitations under the License.
  */
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:velocity_x/src/flutter/selectable_text.dart';