+++ test/flutter/drop_down_button_test.dart
@@ -16,7 +16,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 void main() {
-
   testWidgets('user is able to select value from drop down',
       (WidgetTester tester) async {
     String _value = 'HEY';
@@ -25,8 +24,12 @@ void main() {
       child: MaterialApp(
         home: Material(
           child: ['HELLO', 'HEY', 'HOLA']
-              .dropDown(value: _value)
-              .onChange((value) => _value = value)
+              .textDropDown(
+                selectedValue: _value,
+                onChanged: (value) {
+                  _value = value;
+                },
+              )
               .make()
               .centered(),
         ),